% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/91_data.R
\docType{data}
\name{cigarettes}
\alias{cigarettes}
\title{Cigarette demand}
\format{
A \code{data.frame} object.
}
\usage{
cigarettes
}
\description{
Panel dataset on cigarette demand in 46 US states from 1963 until 1992.
}
\references{
Baltagi, B. H. and Levin, D. (1992). Cigarette taxation: raising revenues
and reducing consumption, \emph{Structural Change and Economic Dynamics},
\bold{3(2)}, 321-335. \doi{10.1016/0954-349X(92)90010-4}.
Baltagi, B. H. and Li, D. (2004). Prediction in the panel data model with
spatial correlation. \emph{Advances in Spatial Econometrics}, 283-295.
Springer, Berlin. \doi{10.1007/978-3-662-05617-2_13}.
}
\keyword{datasets}
