% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_execute.R
\name{sample}
\alias{sample}
\title{Obtain draws from a Bayesian model sampler}
\usage{
sample(x, n_save = 1000L, n_burn = 0L, mh = set_mh(), verbose = TRUE)
}
\arguments{
\item{x}{Bayesian model}

\item{n_save, n_burn}{Integer scalar with number of draws to save / burn.}

\item{mh}{Settings to tune the Metropolis-Hastings step. See \code{\link{set_mh}}.}

\item{verbose}{Logical scalar. Whether to print status updates.}
}
\value{
Returns a numeric matrix with stored draws. The Bayesian model is modified in place.
}
\description{
Obtain draws from a Bayesian model sampler
}
