% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_execute.R
\name{tune}
\alias{tune}
\alias{burn}
\title{Burn-in and tune a Bayesian model sampler}
\usage{
tune(x, n_burn = 1000L, mh = set_mh(), verbose = TRUE)

burn(x, n_burn = 1000L, verbose = TRUE)
}
\arguments{
\item{x}{Bayesian model}

\item{n_burn}{Integer scalar with number of draws to save / burn.}

\item{mh}{Settings to tune the Metropolis-Hastings step. See \code{\link{set_mh}}.}

\item{verbose}{Logical scalar. Whether to print status updates.}
}
\value{
Modifies the Bayesian model in place and returns it invisibly.
}
\description{
Burn-in and tune a Bayesian model sampler
}
