% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/91_data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{United States Historical States}
\format{
A \code{data.frame} object.
}
\source{
\url{https://publications.newberry.org/ahcbp/}
}
\usage{
us_states
}
\description{
Polygons of US state boundaries for the period 1960--2000. Subset from
Siczewicz (2011). Licensed under CC BY-NC-SA 2.5 by the Atlas of Historical
County Boundaries.
}
\references{
Siczewicz, P. (2011) U.S. Historical States and Territories (Generalized
0.01 deg). Dataset. Atlas of Historical County Boundaries, edited by
Long, J. H. . Chicago: The Newberry Library, 2011. Available online from
\url{https://publications.newberry.org/ahcbp/}.
}
\keyword{datasets}
